requirements.txt
# GamblerHunter V2 - 赌徒猎手系统
# 依赖包列表

# 核心依赖
aiohttp>=3.9.0
websockets>=12.0
python-dotenv>=1.0.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# Redis（可选，用于温存储）
redis>=5.0.0

# 高性能文件存储（可选，用于冷存储）
pyarrow>=14.0.0
tables>=3.9.0  # 用于 HDF5 支持

# 异步支持
asyncio-throttle>=1.0.0

# 日志
colorlog>=6.8.0

# 数据验证
pydantic>=2.0.0
